+++ utils.h
@@ -26,6 +26,6 @@ int sizeFromChar(int c);
 
 long longFromHex(char *h);
 long longFromBin(char *h);
-long getRegId(char *reg);
+int getRegId(char *reg);
 void lowercase(char *str);
 int sameSizeRegs(int r1, int r2);
